Remove old about code
