--- app/src/main/res/layout/about_header_item.xml
@@ -1,62 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!--
-  ~ Copyright (c) 2018 Zac Sweers
-  ~
-  ~ Licensed under the Apache License, Version 2.0 (the "License");
-  ~ you may not use this file except in compliance with the License.
-  ~ You may obtain a copy of the License at
-  ~
-  ~   http://www.apache.org/licenses/LICENSE-2.0
-  ~
-  ~ Unless required by applicable law or agreed to in writing, software
-  ~ distributed under the License is distributed on an "AS IS" BASIS,
-  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  ~ See the License for the specific language governing permissions and
-  ~ limitations under the License.
-  -->
-
-<androidx.constraintlayout.widget.ConstraintLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/container"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    android:background="@color/headerBackground"
-    android:clickable="false"
-    android:clipChildren="false"
-    android:focusable="false"
-    android:padding="8dp"
-    >
-
-  <ImageView
-      android:id="@+id/icon"
-      android:layout_width="40dp"
-      android:layout_height="40dp"
-      android:contentDescription="Owner Avatar"
-      android:visibility="visible"
-      app:layout_constraintBottom_toBottomOf="parent"
-      app:layout_constraintEnd_toStartOf="@+id/title"
-      app:layout_constraintStart_toStartOf="parent"
-      app:layout_constraintTop_toTopOf="parent"
-      tools:src="@mipmap/ic_launcher"
-      />
-
-  <io.sweers.catchup.base.ui.widget.BaselineGridTextView
-      android:id="@+id/title"
-      android:layout_width="0dp"
-      android:layout_height="wrap_content"
-      android:layout_marginStart="16dp"
-      android:fontFamily="@font/nunito"
-      android:orientation="vertical"
-      android:textStyle="bold"
-      app:fontFamily="@font/nunito"
-      app:layout_constraintBottom_toBottomOf="parent"
-      app:layout_constraintEnd_toEndOf="parent"
-      app:layout_constraintStart_toEndOf="@+id/icon"
-      app:layout_constraintTop_toTopOf="parent"
-      tools:text="Android"
-      style="@style/TextAppearance.AppCompat.Small"
-      />
-
-</androidx.constraintlayout.widget.ConstraintLayout>
