--- app/src/main/res/layout/fragment_about.xml
@@ -1,110 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?><!--
-  ~ Copyright (c) 2018 Zac Sweers
-  ~
-  ~ Licensed under the Apache License, Version 2.0 (the "License");
-  ~ you may not use this file except in compliance with the License.
-  ~ You may obtain a copy of the License at
-  ~
-  ~   http://www.apache.org/licenses/LICENSE-2.0
-  ~
-  ~ Unless required by applicable law or agreed to in writing, software
-  ~ distributed under the License is distributed on an "AS IS" BASIS,
-  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  ~ See the License for the specific language governing permissions and
-  ~ limitations under the License.
-  -->
-
-<androidx.coordinatorlayout.widget.CoordinatorLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/about_fragment_root"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical"
-    >
-
-  <com.google.android.material.appbar.AppBarLayout
-      android:id="@+id/appbarlayout"
-      android:layout_width="match_parent"
-      android:layout_height="wrap_content"
-      >
-
-    <com.google.android.material.appbar.CollapsingToolbarLayout
-        android:id="@+id/ctl"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        app:layout_scrollFlags="scroll|exitUntilCollapsed"
-        >
-
-      <LinearLayout
-          android:id="@+id/banner_container"
-          android:layout_width="match_parent"
-          android:layout_height="wrap_content"
-          android:orientation="vertical"
-          android:padding="48dp"
-          app:layout_collapseMode="parallax"
-          >
-
-        <ImageView
-            android:id="@+id/banner_icon"
-            android:layout_width="48dp"
-            android:layout_height="48dp"
-            android:layout_gravity="center_horizontal"
-            android:layout_marginBottom="8dp"
-            android:contentDescription="CatchUp icon"
-            android:src="@mipmap/ic_launcher"
-            />
-
-        <TextView
-            android:id="@+id/banner_title"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_gravity="center_horizontal"
-            android:layout_marginBottom="8dp"
-            android:gravity="center"
-            android:text="@string/app_name"
-            android:textAppearance="@style/TextAppearance.AppCompat.Title"
-            />
-
-        <io.sweers.catchup.base.ui.widget.BaselineGridTextView
-            android:id="@+id/banner_text"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_gravity="center_horizontal"
-            android:gravity="center"
-            android:textAppearance="@style/TextAppearance.AppCompat.Small"
-            android:textColorLink="?android:textColorPrimary"
-            tools:text="An app for catching up on things.\n\nv1.0\nBy Zac Sweers  —  Source code"
-            />
-      </LinearLayout>
-
-      <androidx.appcompat.widget.Toolbar
-          android:id="@+id/toolbar"
-          android:layout_width="wrap_content"
-          android:layout_height="?actionBarSize"
-          app:layout_collapseMode="pin"
-          app:titleMarginStart="72dp"
-          />
-    </com.google.android.material.appbar.CollapsingToolbarLayout>
-
-    <com.google.android.material.tabs.TabLayout
-        android:id="@+id/tab_layout"
-        style="@style/CatchUp.TabLayout.Inverted"
-        android:layout_width="match_parent"
-        android:layout_height="?actionBarSize"
-        android:background="@null"
-        app:tabMode="fixed"
-        app:tabUnboundedRipple="false"
-        />
-
-  </com.google.android.material.appbar.AppBarLayout>
-
-  <androidx.viewpager2.widget.ViewPager2
-      android:id="@+id/view_pager"
-      android:layout_width="match_parent"
-      android:layout_height="wrap_content"
-      android:overScrollMode="never"
-      app:layout_behavior="@string/appbar_scrolling_view_behavior"
-      />
-
-</androidx.coordinatorlayout.widget.CoordinatorLayout>
